refactor(agent): separate control and operation domains
